+++ crates/full-node/sov-sequencer/src/standard/mod.rs
@@ -80,7 +80,8 @@ where
     inner: Mutex<Inner<S, Rt, Da>>,
     checkpoint_sender: watch::Sender<StateCheckpoint<S>>,
     api_state: ApiState<S>,
-    config: SequencerConfig<S::Da, S::Address, StdSequencerConfig>,
+    da_address: <S::Da as DaSpec>::Address,
+    config: SequencerConfig<S::Address, StdSequencerConfig>,
     api_ledger_db: LedgerDb,
 }
 
@@ -111,7 +112,7 @@ where
         state_update_receiver: StateUpdateReceiver<S::Storage>,
         _da_sync_state: Arc<DaSyncState>,
         storage_path: &Path,
-        config: &SequencerConfig<S::Da, S::Address, StdSequencerConfig>,
+        config: &SequencerConfig<S::Address, StdSequencerConfig>,
         ledger_db: LedgerDb,
         api_ledger_db: LedgerDb,
         shutdown_sender: watch::Sender<()>,
@@ -136,6 +137,7 @@ where
         let checkpoint =
             StateCheckpoint::new(latest_state_update.storage.clone(), &runtime.kernel());
 
+        let da_address = da.get_signer().await;
         let (blob_sender, blob_sender_handle) = BlobSender::new(
             da,
             ledger_db.clone(),
@@ -173,6 +175,7 @@ where
             checkpoint_sender,
             config: config.clone(),
             api_ledger_db,
+            da_address,
         });
 
         handles.push(tokio::spawn({
@@ -244,7 +247,7 @@ where
             &self.runtime,
             message,
             &authz_data.default_address,
-            &self.config.da_address,
+            &self.da_address,
             &self.config.admin_addresses,
         ) {
             ctx.state_checkpoint = tx_scratchpad.revert();
@@ -264,7 +267,7 @@ where
             &<S::Gas>::MAX,
             auth_output,
             mempool_tx.tx.clone(),
-            &self.config.da_address,
+            &self.da_address,
             self.config.rollup_address.clone(),
             ExecutionContext::Sequencer,
             &NoOpControlFlow,