+++ crates/module-system/sov-test-utils/src/sequencer.rs
@@ -37,7 +37,7 @@ use crate::{
 pub struct TestSequencerSetup<Rt: Runtime<TestSpec>> {
     _dir: TempDir,
     /// The [`SequencerConfig`] used in this test.
-    pub config: SequencerConfig<MockDaSpec, <TestSpec as Spec>::Address, StdSequencerConfig>,
+    pub config: SequencerConfig<<TestSpec as Spec>::Address, StdSequencerConfig>,
     /// The DA service used by the sequencer.
     pub da_service: StorableMockDaService,
     /// What was passed to Sequencer::create.
@@ -145,7 +145,6 @@ impl<Rt: Runtime<TestSpec>> TestSequencerSetup<Rt> {
         shutdown_receiver.mark_unchanged();
 
         let config = SequencerConfig {
-            da_address: da_service.sequencer_da_address,
             rollup_address: sequencer_rollup_address,
             admin_addresses,
             automatic_batch_production: true,