+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2025-07-21
 - #3263 **BREAKING CHANGE** Update error object used for errors returned by REST API endpoints.
     Previously `{errors: [{$ERR_OBJ}], meta: {}}`, now we just return a single `{$ERR_OBJ}` directly.
+- #3279 **BREAKING CHANGE** `sequencer.da_address` is removed from rollup_config.toml. 
 - #3271 adds `sov-zkvm-utils` crate for making ZKVM build.rs smaller
 
 # 2025-07-19